test-iot.bbclass: support manifest file created for swupd images
At the moment, swupd images have a .rootfs.manifest file
as specified by the IMAGE_MANIFEST variable, but not the
symlink to it that test-iot.bbclass was relying upon.

Even if the symlink gets added, using the pre-defined
IMAGE_MANIFEST variable is simpler (no need to compose
the full path manually) and has the same effect.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>